+++ app/controllers/application_controller.rb
@@ -1,12 +1,5 @@
 class ApplicationController < ActionController::API
   include ActionController::ImplicitRender
-
-  include SkipTrackable
-  include TokenAuthentication
-
-  decent_configuration do
-    strategy DecentExposure::StrongParametersStrategy
-  end
-
+  acts_as_token_authentication_handler_for User
   respond_to :json
 end