+++ app/controllers/v1/sessions_controller.rb
@@ -2,11 +2,9 @@ module V1
   class SessionsController < Devise::SessionsController
     wrap_parameters :user
 
-    skip_before_action :skip_trackable
-
     def create
       user = AuthenticateUser.call(warden: warden).user
-      respond_with(user)
+      respond_with(user, serializer: SessionSerializer)
     end
   end
 end