+++ app/models/user.rb
@@ -1,21 +1,8 @@
 class User < ActiveRecord::Base
-  devise :database_authenticatable, :registerable,
-         :recoverable, :trackable, :validatable
+  acts_as_token_authenticatable
 
-  before_save :ensure_authentication_token!
+  devise :database_authenticatable, :registerable,
+    :recoverable, :trackable, :validatable
 
   has_many :comments
-
-  def ensure_authentication_token!
-    self.authentication_token = generate_authentication_token if authentication_token.blank?
-  end
-
-  private
-
-  def generate_authentication_token
-    loop do
-      token = Devise.friendly_token
-      break token unless User.find_by_authentication_token(token)
-    end
-  end
 end