+++ app/serializers/session_serializer.rb
@@ -1,0 +1,3 @@
+class SessionSerializer < UserSerializer
+  attributes :authentication_token
+end