+++ config/application.rb
@@ -1,9 +1,9 @@
-require File.expand_path('../boot', __FILE__)
+require File.expand_path("../boot", __FILE__)
 
 # Pick the frameworks you want:
-require 'active_record/railtie'
-require 'action_controller/railtie'
-require 'action_mailer/railtie'
+require "active_record/railtie"
+require "action_controller/railtie"
+require "action_mailer/railtie"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -24,10 +24,12 @@ module RailsBaseApi
     # config.i18n.default_locale = :de
 
     # Default e-mail address which will be shown in the "from" devise emails, initializers/devise.rb.
-    config.noreply = 'noreply@fs-rails-base-api.heroku.com'
+    config.noreply = "noreply@fs-rails-base-api.heroku.com"
 
     # Default host for action mailer, initializers/mailer.rb
-    config.host = 'localhost:5000'
+    config.host = "localhost:5000"
+
+    config.serve_static_files = false
 
     # Disable default Rails headers which do not make sense in
     # API-only project (X-Frame-Options, X-XSS-Protection, X-Content-Type-Options)