+++ config/environments/development.rb
@@ -27,4 +27,5 @@ Rails.application.configure do
 
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
+  config.serve_static_files = true
 end