+++ spec/support/matchers/be_a_user_representation.rb
@@ -2,7 +2,6 @@ RSpec::Matchers.define :be_a_user_representation do |user|
   match do |json|
     response_attributes = user.sliced_attributes %w(
       id
-      authentication_token
       email
     )